+++ src/main/java/lotto/util/StringUtil.java
@@ -1,0 +1,11 @@
+package lotto.util;
+
+public class StringUtil {
+    private static final String SEPARATOR = ",";
+
+    public static String[] split(String str) {
+        return str.split(SEPARATOR);
+    }
+
+
+}
